+++ R/spFromTable.r
@@ -44,8 +44,11 @@ spFromTable <- function(x, attr_tab =  NULL, crs, ..., topol_ = NULL) {
   if (missing(crs)) crs <- attr(x, "crs")
   if (is.null(crs)) crs <- NA_character_
   ## raster::geom form
-  if (is.null(topol_)) target <- detectSpClass(x)
-  dat <- x %>% distinct_("object_", .keep_all = TRUE)
+  if (is.null(topol_)) target <- detectSpClass(x) else target <- topol_
+  ## check for minimum sensible number of coordinates
+  minc <- c(SpatialPolygonsDataFrame = 3, SpatialLinesDataFrame = 2, SpatialMultiPointsDataFrame = 1, SpatialPointsDataFrame = 1)[target]
+  if (nrow(x) < minc) stop(sprintf("target is %s but input table has  %i %s", target, nrow(x), c("rows", "row")[(nrow(x) ==1)+1]))
+  dat <- distinct_(x, "object_", .keep_all = TRUE)
 
    n_object <- nrow(dat)
    n_attribute <- nrow(attr_tab)
@@ -83,17 +86,18 @@ reverse_geomPoly <- function(x, d, proj) {
   ## match.ID should be replaced by method to carry the original rownames somehow
   SpatialPolygonsDataFrame(SpatialPolygons(lapply(objects, loopBranchPoly), proj4string = CRS(proj)), d, match.ID = FALSE)
 }
-loopBranchPoly <- function(a) Polygons(lapply(dropZeroRowFromList(split(a, a$branch_)), function(b) Polygon(as.matrix(b[, c("x_", "y_")]), hole = !b$island_[1L] == 1)), as.character(a$object_[1L]))
+loopBranchPoly <- function(a) Polygons(lapply(dropZeroRowFromList(split(a, a$branch_)), 
+                                              function(b) Polygon(as.matrix(b[, c("x_", "y_")]), hole = !b$island_[1L])), as.character(a$object_[1L]))
 
 
 reverse_geomLine <- function(x, d, proj) {
   objects <- split(x, x$object_)
   d$branch_ <- d$object_ <- d$order_ <- d$x_ <- d$y_ <- NULL
   if (ncol(d) < 1L) d$rownumber_ <- seq(nrow(d))  ## we might end up with no attributes
-  SpatialLinesDataFrame(SpatialLines(lapply(objects, loopBranchLine), proj4string = CRS(proj)), d)
+  SpatialLinesDataFrame(SpatialLines(lapply(objects, loopBranchLine), proj4string = CRS(proj)), d, match.ID = FALSE)
 }
-dropZeroRowFromList <- function(x) x[unlist(lapply(x, nrow)) > 0L]
-loopBranchLine<- function(a) Lines(lapply(dropZeroRowFromList(split(a, a$branch_)), function(b) Polygon(as.matrix(b[, c("x_", "y_")]))), as.character(a$object_[1L]))
+dropZeroRowFromList <- function(x) x[unlist(lapply(x, nrow), use.names = FALSE) > 0L]
+loopBranchLine<- function(a) Lines(lapply(dropZeroRowFromList(split(a, a$branch_)), function(b) Line(as.matrix(b[, c("x_", "y_")]))), as.character(a$object_[1L]))
 
 reverse_geomPoint <- function(a, d, proj) {
   # stop("not implemented")
@@ -123,7 +127,7 @@ detectSpClass <- function(x) {
   #if (all(gn$SpatialPointsDataFrame %in% names(x))) return("SpatialPointsDataFrame")
   #if (all(gn$SpatialMultiPointsDataFrame %in% names(x))) return("SpatialMultiPointsDataFrame")
   cat("cannot find matching topology type from these columns")
-  print(x)
+  #print(x)
   stop('cannot create Spatial* object from this input')
   
 }